# swagger.yaml
openapi: 3.0.3
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:7070
    description: Generated server url
paths:
  /versions/candidates/{versionCandidateId}/forms/{formName}:
    get:
      tags:
        - Candidate version forms Rest API
      summary: Get form
      operationId: getForm_1
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
        - name: versionCandidateId
          in: path
          required: true
          schema:
            type: string
        - name: formName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
    put:
      tags:
        - Candidate version forms Rest API
      summary: Update form
      operationId: updateForm
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
        - name: versionCandidateId
          in: path
          required: true
          schema:
            type: string
        - name: formName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '201':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
    post:
      tags:
        - Candidate version forms Rest API
      summary: Create new form
      operationId: formCreate
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
        - name: versionCandidateId
          in: path
          required: true
          schema:
            type: string
        - name: formName
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        required: true
      responses:
        '201':
          description: OK
          content:
            application/json: {}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '422':
          description: Unprocessable Entity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
    delete:
      tags:
        - Candidate version forms Rest API
      summary: Delete form
      operationId: deleteForm
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
        - name: versionCandidateId
          in: path
          required: true
          schema:
            type: string
        - name: formName
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: OK
          content:
            '*/*':
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
...
  /versions/master/forms:
    get:
      tags:
        - Master version forms Rest API
      summary: Get forms list for master version
      operationId: getFormsFromMaster
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShortFormDetails'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
  /versions/master/forms/{formName}:
    get:
      tags:
        - Master version forms Rest API
      summary: Get form
      operationId: getForm
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
        - name: formName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
  /versions/master/forms/{formName}/download:
    get:
      tags:
        - Master version forms Rest API
      summary: Download form
      operationId: downloadForm
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
        - name: formName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/octet-stream: {}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
  /versions/master/:
    get:
      tags:
        - Master version Rest API
      summary: Get master version
      operationId: getVersionDetails
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MasterVersionResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
  /versions/candidates/{versionCandidateId}:
    get:
      tags:
        - Candidate version Rest API
      summary: Get version details by id
      operationId: getVersionDetails_1
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
        - name: versionCandidateId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VersionDetailedResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
...
  /versions/candidates/{versionCandidateId}/forms/{formName}/download:
    get:
      tags:
        - Candidate version forms Rest API
      summary: Download form
      operationId: downloadForm_1
      parameters:
        - name: X-Access-Token
          in: header
          schema:
            type: string
        - name: versionCandidateId
          in: path
          required: true
          schema:
            type: string
        - name: formName
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/octet-stream: {}
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DetailedErrorResponse'
...
components:
  schemas:
    DetailedErrorResponse:
      type: object
      properties:
        traceId:
          type: string
        code:
          type: string
        details:
          type: string
        localizedMessage:
          type: string
...
    ShortFormDetails:
      type: object
      properties:
        name:
          type: string
        title:
          type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
    MasterVersionResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        author:
          type: string
        latestUpdate:
          type: string
          format: date-time
        published:
          type: boolean
        inspector:
          type: string
        validations:
          type: array
          items:
            $ref: '#/components/schemas/Validation'
    Validation:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - REGULATION_INTEGRITY
            - TEST
        result:
          type: string
          enum:
            - APPROVED
            - REJECTED
        resultDetails:
          type: string
...
    Inspection:
      type: object
      properties:
        name:
          type: string
        inspector:
          type: string
        result:
          type: string
          enum:
            - APPROVED
            - REJECTED
        resultDetails:
          type: string
    VersionDetailedResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        author:
          type: string
        creationDate:
          type: string
          format: date-time
        latestUpdate:
          type: string
          format: date-time
        published:
          type: boolean
        hasConflicts:
          type: boolean
        inspections:
          type: array
          items:
            $ref: '#/components/schemas/Inspection'
        validations:
          type: array
          items:
            $ref: '#/components/schemas/Validation'